+++ src/components/molecules/ProtocolsIcons/styles.scss
@@ -1,0 +1,5 @@
+.protocols-icons {
+  & img:not(:first-child) {
+    margin-left: -4px;
+  }
+}
