+++ src/components/molecules/index.tsx
@@ -5,6 +5,7 @@ export { default as CrossChainChart } from "./CrossChainChart";
 export { default as ErrorPlaceholder } from "./ErrorPlaceholder";
 export { default as Footer } from "./Footer";
 export { default as Header } from "./Header";
+export { default as ProtocolsIcons } from "./ProtocolsIcons";
 export { default as ProtocolsStats } from "./ProtocolsStats";
 export { default as StatusBadge } from "./StatusBadge";
 export { default as TermsOfUseBanner } from "./TermsOfUseBanner";
