+++ src/pages/Txs/Information/styles.scss
@@ -178,12 +178,6 @@
         }
       }
 
-      .tx-protocol-icons {
-        & img:not(:first-child) {
-          margin-left: -4px;
-        }
-      }
-
       .tx-view-details {
         & > a {
           @include button-secondary;
